Refaktorer: flytt hovedmålfilter til eneste sted det blir brukt
