+++ src/veileder/sok/hovedmal/Hovedmal.less
